Fix mason installs and auto-formatting
LSPs and formatters were not correctly installing due to lazy-loading
(classic...). Also, vim format was invoking both none-ls and LSP
formatters in an unspecified order, which caused formatting
discrepancies.

To fix the former, I refactored the installation of related plugins.

To fix the latter, I migrated to conform.nvim. It's really good!\

I also applied formatting to all files in this commit, and added a
stylua.toml.